test(#2429): batch 16 — in-process closure: drain 740/993 + 517/42/52/94 halves (143 items)
- 740 (88): 60 unit + 28 routes — the 24 ambient-PG route bootstraps
  (_make_client/create_app) repaired to the canonical tmp-sqlite fixtures;
  per-test ledger in the plan; 2 playwright scripts -> tests/e2e
- 517 in-process half (27 -> 14 after audited dedupe: 5 subsumed
  deletions + 1 disjoint-merge, per-def ledger); server/browser-bound
  siblings stay byte-stable for batch 17
- 42 webui_manager (9): single-user launch stubbed; stale port assertions
  realigned to the post-#3129 3100 pin (masked by the launch ValueError)
- 52 (4): bool-return bodies -> real asserts; 94 (5): zero-assert
  data-audit scripts -> seeded sqlite + real coverage asserts via the
  _db_url_cache pin; 993 (10) straight move + script rehomed
- 6 no-assert __main__ scripts deleted (32x2, 69x1, 51x3 — 0 items)
- 3 governance rows; scanner: 2 should-not-raise strengthenings + legacy
  runner dropped (69 = main)
- Baselines: issues 99->169(clamped mirror, delta negative)/110->93
  (12 migrated + 5 deleted globs); default min 7975->8112 (+137);
  floor >=93; inventory 63->61. Failure-set diff vs base: 0 new.
