+++ .test-baseline.json
@@ -4,12 +4,12 @@
   "created_before_modification": true,
   "layers": {
     "default": {
-      "min_tests": 7975,
+      "min_tests": 8112,
       "min_files": 250,
       "last_updated": "2026-08-27",
-      "update_reason": "Issue #2429 batch 15: B-group wave 2 (dirs 716+610) migrated 500 items (355 unit, 42 integration flat, 103 integration/routes); min raised by exactly +500 (7475+500); actuals measured locally, CI clean re-confirmed",
-      "actual_collected": 11609,
-      "actual_files": 769,
+      "update_reason": "Issue #2429 batch 16: in-process closure migrated 137 items (98 unit, 34 routes, 5 integration); min raised by exactly +137 (7975+137); actuals re-measured on branch (base drifted +29 via parallel PRs since batch 15's record)",
+      "actual_collected": 11779,
+      "actual_files": 784,
       "threshold_percent": 96
     },
     "critical": {
@@ -30,12 +30,12 @@
       "note": "Currently excluded by norecursedirs"
     },
     "issues": {
-      "min_tests": 99,
-      "min_files": 110,
+      "min_tests": 169,
+      "min_files": 93,
       "last_updated": "2026-08-27",
-      "update_reason": "Issue #2429 batch 15: B-group wave 2 \u2014 dirs 716/610, 500 items / 16 glob files, all migrated (repairs: 716 scheduler conf-seam pin + agent_runner #2019 proxy stub + autonomous_api shared-checkout env pin, 610 git_endpoints rewritten to real route import killing the MagicMock __spec__ import-machinery class; isolation fixes: cross-file workflow rate-limiter state + remote-agent sys.modules config leak); 2 standalone playwright scripts rehomed to tests/e2e + governance rows (+2, sanctioned hand-edit; validate 0 errors; governance tests 45/45)",
-      "actual_collected": 312,
-      "actual_files": 125,
+      "update_reason": "Issue #2429 batch 16: in-process closure \u2014 143 items drained (137 migrated + 6 audited 517 cross-file dedupe subsumptions with per-def ledger), 12 glob files: dirs 740 (88; 24 ambient-PG route bootstraps repaired to canonical tmp-sqlite fixtures) and 993 (10 + script) fully drained; 517 in-process half (27 incl. dedupe), 42 webui_manager (9; stale port assertions realigned to post-#3129 3100 pin), 52 (4 bool-returns -> asserts), 94 (5 zero-assert data-audit scripts -> seeded sqlite + real coverage asserts) partially drained; 6 no-assert __main__ scripts deleted (32x2, 69x1, 51x3); 3 playwright scripts rehomed to tests/e2e + governance rows; min_tests delta would be 99-143 < 0 -> CLAMPED to mirror remaining clean count 169 (protective for the final pre-retirement window); min_files 110-17=93 exact (12 migrated globs + 5 deleted no-assert script globs)",
+      "actual_collected": 169,
+      "actual_files": 108,
       "threshold_percent": 97
     }
   },
@@ -47,7 +47,7 @@
     "Baseline created before pytest.ini modification to enable e2e_*.py collection",
     "tests/e2e currently excluded by norecursedirs in pytest.ini",
     "Issue item baseline is enforced by ci.yml; the extended runner scales the file baseline per shard",
-    "Default count from: python scripts/ci.py run default-collection (11609 items after #2429 batch 15)",
-    "Issue count from: clean-checkout issue-collection (312 items, 125 files after #2429 batch 15; local dev venv may read 313/126 when playwright is installed \u2014 dir 41 importorskip)"
+    "Default count from: python scripts/ci.py run default-collection (11775 items after #2429 batch 16)",
+    "Issue count from: clean-checkout issue-collection (169 items, 130 files after #2429 batch 16)"
   ]
 }
