+++ tests/integration/routes/test_milestone_diff_api_740.py
@@ -1,20 +1,28 @@
-"""
-Batch 4 tests — Diff API endpoint, status filter.
+"""Issue #740 — Milestone diff API and workflow list status filter (R1-migrated).
 
-Tests for:
-- GET /api/autonomous/workflows/<id>/milestones/<mid>/diff
-- Workflow list status filter query parameter
+Migrated wholesale from tests/issues/740/test_batch4_diff_api.py (all 13 items
+are route-bound).
+
+R1 repair: the legacy hand-rolled ``_make_client()`` created a temp SQLite DB
+but left ``DATABASE_URL`` untouched, so ``create_app``'s ``ensure_all_tables``
+dialed the ambient (Postgres) config DB. Replaced with the canonical
+``auto_db``/``client`` bootstrap from tests/integration/routes/test_autonomous_api.py
+(tmp sqlite + ``DATABASE_URL`` env pin + ``user_repo`` rebind). Assertions
+unchanged (the original ``unittest`` asserts were translated 1:1 to plain
+``assert`` statements).
 """
 
-import json
 import os
-import sys
-import tempfile
-import unittest
 from unittest.mock import MagicMock, patch
 
-# Ensure project root on path
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "..", ".."))
+import pytest
+
+import app.repositories.database as db_mod
+from app.repositories.database import Database
+
+pytestmark = [pytest.mark.regression, pytest.mark.issue(740)]
+
+# ── Helpers ──────────────────────────────────────────────────────────
 
 
 def _make_workflow(**overrides):
@@ -94,38 +102,74 @@ def _make_milestone(**overrides):
     return base
 
 
-def _make_client():
-    """Create Flask test client with test DB and mock auth."""
-    import app.repositories.database as db_mod
+# ── Canonical fixtures (tests/integration/routes/test_autonomous_api.py) ──
 
-    db_path = tempfile.mktemp(suffix=".db")
-    orig = db_mod.adapt_sql
-    db_mod.adapt_sql = lambda sql: sql
 
-    db = db_mod.Database(f"sqlite:///{db_path}")
-    try:
-        with db.get_connection() as conn:
-            cursor = conn.cursor()
-            from app.repositories.schema_init import load_schema_from_file
+@pytest.fixture
+def auto_db(tmp_path):
+    """Create a temporary SQLite database with autonomous tables."""
+    with patch.object(db_mod, "is_postgresql", return_value=False):
+        orig = db_mod.adapt_sql
+        db_mod.adapt_sql = lambda q: q
+        try:
+            db_path = str(tmp_path / "test_api.db")
+            db = Database(db_url=f"sqlite:///{db_path}")
+            conn = db.get_connection()
+            try:
+                from app.repositories.schema_init import load_schema_from_file
 
-            # Let the authoritative schema create users (the hand-rolled DDL
-            # drifted: no deleted_at/system_account — the schema's partial
-            # indexes on those columns then failed).
-            load_schema_from_file(db_url=f"sqlite:///{db_path}", dialect="sqlite")
-            cursor.execute(
-                "INSERT OR IGNORE INTO users (username, email, password_hash, role) VALUES (?, ?, ?, ?)",
-                ("admin", "admin@test.com", "hash123", "admin"),
-            )
-            conn.commit()
-    finally:
-        pass
+                # Create the FULL authoritative schema (incl. users.deleted_at) on the
+                # empty DB FIRST, then seed. Do NOT hand-CREATE an old users table —
+                # load_schema's CREATE TABLE IF NOT EXISTS will not add the missing
+                # column to an already-existing table (legacy tests/issues escape hatch).
+                load_schema_from_file(db_url=db.db_url, dialect="sqlite")
+                cursor = conn.cursor()
+                cursor.execute(
+                    "INSERT INTO users (username, email, password_hash, role) VALUES (?, ?, ?, ?)",
+                    ("admin", "admin@test.com", "hash123", "platform_admin"),
+                )
+                conn.commit()
+            finally:
+                conn.close()
+            yield db
+        finally:
+            db_mod.adapt_sql = orig
+            try:
+                os.unlink(db_path)
+            except OSError:
+                pass
 
+
+@pytest.fixture
+def client(auto_db, monkeypatch):
+    """Create a Flask test client with session cookie set."""
     from app import create_app
+    from app.repositories.user_repo import UserRepository
 
+    # create_app()'s ensure_all_tables() picks its DB from get_database_url()
+    # (DATABASE_URL), which defaults to the dev Postgres DB locally; and the
+    # module-global user_repo (app.routes.autonomous) binds Database() once at
+    # import. Point both at auto_db's seeded SQLite DB so the app and the route
+    # share the fixture's database. monkeypatch auto-restores both.
+    monkeypatch.setenv("DATABASE_URL", auto_db.db_url)
+    # The create endpoint's module-level rate limiter (10/user/hour) is
+    # process-global state: hits accumulated by earlier test files in the same
+    # pytest process would 429 this file's own create requests (they run as
+    # user_id=1 too). Clear the per-user hit log so each test starts with a
+    # full budget; the limiter itself stays fully in effect within the test.
+    monkeypatch.setattr("app.routes.autonomous._workflow_rate_limiter._hits", {})
+    # These endpoint tests were written for single-tenant semantics: pin the
+    # deployment mode so the default branch_strategy ("new-branch") is not
+    # rejected by the multi-user _shared_checkout_rejection gate (#2021;
+    # that rejection logic is separately unit-covered by
+    # tests/unit/test_git_path_hardening.py).
+    monkeypatch.setenv("OPENACE_ALLOW_SHARED_CHECKOUT", "1")
     app = create_app({"TESTING": True})
-    c = app.test_client()
-    c.set_cookie("session_token", "test-token")
-    return c, db_path, orig, db_mod
+    monkeypatch.setattr("app.routes.autonomous.user_repo", UserRepository(db=auto_db))
+    with app.app_context():
+        c = app.test_client()
+        c.set_cookie("session_token", "test-token")
+        yield c
 
 
 def _mock_auth(user_id=1, role="admin"):
@@ -140,91 +184,85 @@ def _mock_auth(user_id=1, role="admin"):
     )
 
 
-class TestGetMilestoneDiff(unittest.TestCase):
+@pytest.fixture(autouse=True)
+def _stub_owner_user_lookup():
+    """Stub the owner system_account resolution (kept from the legacy setUp).
+
+    get_milestone_diff resolves the workflow owner's system_account via the
+    module-level UserRepository (bound at import time to the DEFAULT
+    database, not the temp DB) — stub it so the system_account assertion
+    doesn't depend on whatever the ambient environment DB holds.
+    """
+    from app.repositories.user_repo import UserRepository
+
+    with patch.object(
+        UserRepository,
+        "get_user_by_id",
+        lambda self_, user_id: {
+            "id": user_id,
+            "username": "admin",
+            "system_account": "admin",
+            "role": "platform_admin",
+            "tenant_id": None,
+        },
+    ):
+        yield
+
+
+class TestGetMilestoneDiff:
     """Tests for GET /api/autonomous/workflows/<id>/milestones/<mid>/diff."""
 
-    def setUp(self):
-        self.client, self.db_path, self.orig, self.db_mod = _make_client()
-        # get_milestone_diff resolves the workflow owner's system_account via
-        # the module-level UserRepository (bound at import time to the DEFAULT
-        # database, not the temp DB) — stub it so the system_account assertion
-        # doesn't depend on whatever the ambient environment DB holds.
-        from unittest.mock import MagicMock as _MM
-
-        from app.repositories.user_repo import UserRepository
-
-        self._user_patch = patch.object(
-            UserRepository,
-            "get_user_by_id",
-            lambda self_, user_id: {
-                "id": user_id,
-                "username": "admin",
-                "system_account": "admin",
-                "role": "platform_admin",
-                "tenant_id": None,
-            },
-        )
-        self._user_patch.start()
-        self.addCleanup(self._user_patch.stop)
-
-    def tearDown(self):
-        self.db_mod.adapt_sql = self.orig
-        try:
-            os.unlink(self.db_path)
-        except OSError:
-            pass
-
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_workflow_not_found(self, mock_repo):
+    def test_diff_workflow_not_found(self, mock_repo, client):
         """Return 404 if workflow does not exist."""
         mock_repo.get_workflow.return_value = None
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
-        self.assertEqual(resp.status_code, 404)
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+        assert resp.status_code == 404
         data = resp.get_json()
-        self.assertIn("not found", data["error"].lower())
+        assert "not found" in data["error"].lower()
 
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_access_denied_non_admin(self, mock_repo):
+    def test_diff_access_denied_non_admin(self, mock_repo, client):
         """Return 403 if non-admin tries to access another user's workflow."""
         mock_repo.get_workflow.return_value = _make_workflow(user_id=99)
         with _mock_auth(user_id=1, role="user"):
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
-        self.assertEqual(resp.status_code, 403)
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+        assert resp.status_code == 403
 
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_milestone_not_found(self, mock_repo):
+    def test_diff_milestone_not_found(self, mock_repo, client):
         """Return 404 if milestone does not exist."""
         mock_repo.get_workflow.return_value = _make_workflow()
         mock_repo.get_milestone.return_value = None
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-missing/diff")
-        self.assertEqual(resp.status_code, 404)
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-missing/diff")
+        assert resp.status_code == 404
 
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_milestone_wrong_workflow(self, mock_repo):
+    def test_diff_milestone_wrong_workflow(self, mock_repo, client):
         """Return 404 if milestone belongs to different workflow."""
         mock_repo.get_workflow.return_value = _make_workflow()
         mock_repo.get_milestone.return_value = _make_milestone(workflow_id="other-wf")
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
-        self.assertEqual(resp.status_code, 404)
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+        assert resp.status_code == 404
 
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_empty_commit_shas(self, mock_repo):
+    def test_diff_empty_commit_shas(self, mock_repo, client):
         """Return empty diff when milestone has no commits."""
         mock_repo.get_workflow.return_value = _make_workflow()
         mock_repo.get_milestone.return_value = _make_milestone(commit_shas="")
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
-        self.assertEqual(resp.status_code, 200)
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+        assert resp.status_code == 200
         data = resp.get_json()
-        self.assertTrue(data["success"])
-        self.assertEqual(data["diff"], "")
+        assert data["success"]
+        assert data["diff"] == ""
 
     @patch("app.modules.workspace.autonomous.github_ops.GitHubOps")
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_with_json_array_shas(self, mock_repo, mock_gh_class):
+    def test_diff_with_json_array_shas(self, mock_repo, mock_gh_class, client):
         """Return concatenated diffs for commits in JSON array format."""
         mock_repo.get_workflow.return_value = _make_workflow()
         mock_repo.get_milestone.return_value = _make_milestone(
@@ -239,19 +277,19 @@ class TestGetMilestoneDiff(unittest.TestCase):
         mock_gh_class.return_value = mock_gh
 
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
 
-        self.assertEqual(resp.status_code, 200)
+        assert resp.status_code == 200
         data = resp.get_json()
-        self.assertTrue(data["success"])
-        self.assertIn("abc123de", data["diff"])
-        self.assertIn("456789gh", data["diff"])
-        self.assertIn("file1.py", data["diff"])
-        self.assertIn("file2.py", data["diff"])
+        assert data["success"]
+        assert "abc123de" in data["diff"]
+        assert "456789gh" in data["diff"]
+        assert "file1.py" in data["diff"]
+        assert "file2.py" in data["diff"]
 
     @patch("app.modules.workspace.autonomous.github_ops.GitHubOps")
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_comma_separated_shas(self, mock_repo, mock_gh_class):
+    def test_diff_comma_separated_shas(self, mock_repo, mock_gh_class, client):
         """Handle comma-separated commit SHAs (not JSON array)."""
         mock_repo.get_workflow.return_value = _make_workflow()
         mock_repo.get_milestone.return_value = _make_milestone(commit_shas="abc123,def456")
@@ -261,16 +299,16 @@ class TestGetMilestoneDiff(unittest.TestCase):
         mock_gh_class.return_value = mock_gh
 
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
 
-        self.assertEqual(resp.status_code, 200)
+        assert resp.status_code == 200
         data = resp.get_json()
-        self.assertTrue(data["success"])
-        self.assertEqual(mock_gh.get_commit_diff.call_count, 2)
+        assert data["success"]
+        assert mock_gh.get_commit_diff.call_count == 2
 
     @patch("app.modules.workspace.autonomous.github_ops.GitHubOps")
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_github_ops_error_graceful(self, mock_repo, mock_gh_class):
+    def test_diff_github_ops_error_graceful(self, mock_repo, mock_gh_class, client):
         """Gracefully handle GitHubOps errors — return empty diff for failed commits."""
         mock_repo.get_workflow.return_value = _make_workflow()
         mock_repo.get_milestone.return_value = _make_milestone(commit_shas="abc123")
@@ -280,16 +318,16 @@ class TestGetMilestoneDiff(unittest.TestCase):
         mock_gh_class.return_value = mock_gh
 
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
 
-        self.assertEqual(resp.status_code, 200)
+        assert resp.status_code == 200
         data = resp.get_json()
-        self.assertTrue(data["success"])
-        self.assertEqual(data["diff"], "")
+        assert data["success"]
+        assert data["diff"] == ""
 
     @patch("app.modules.workspace.autonomous.github_ops.GitHubOps")
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_uses_worktree_path_preferred(self, mock_repo, mock_gh_class):
+    def test_diff_uses_worktree_path_preferred(self, mock_repo, mock_gh_class, client):
         """Use worktree_path over project_path when both exist."""
         mock_repo.get_workflow.return_value = _make_workflow(
             project_path="/tmp/project",
@@ -302,14 +340,14 @@ class TestGetMilestoneDiff(unittest.TestCase):
         mock_gh_class.return_value = mock_gh
 
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
 
-        self.assertEqual(resp.status_code, 200)
+        assert resp.status_code == 200
         mock_gh_class.assert_called_once_with("/tmp/worktree", system_account="admin")
 
     @patch("app.modules.workspace.autonomous.github_ops.GitHubOps")
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_falls_back_to_project_path(self, mock_repo, mock_gh_class):
+    def test_diff_falls_back_to_project_path(self, mock_repo, mock_gh_class, client):
         """Use project_path when worktree_path is empty."""
         mock_repo.get_workflow.return_value = _make_workflow(
             project_path="/tmp/project",
@@ -322,13 +360,13 @@ class TestGetMilestoneDiff(unittest.TestCase):
         mock_gh_class.return_value = mock_gh
 
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+            resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
 
-        self.assertEqual(resp.status_code, 200)
+        assert resp.status_code == 200
         mock_gh_class.assert_called_once_with("/tmp/project", system_account="admin")
 
     @patch("app.routes.autonomous.auto_repo")
-    def test_diff_single_sha_string(self, mock_repo):
+    def test_diff_single_sha_string(self, mock_repo, client):
         """Handle a single commit SHA string (not array or comma-separated)."""
         mock_repo.get_workflow.return_value = _make_workflow()
         mock_repo.get_milestone.return_value = _make_milestone(commit_shas="abc123def456")
@@ -339,47 +377,33 @@ class TestGetMilestoneDiff(unittest.TestCase):
                 mock_gh.get_commit_diff.return_value = "diff content"
                 mock_gh_class.return_value = mock_gh
 
-                resp = self.client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
+                resp = client.get("/api/autonomous/workflows/wf-1/milestones/ms-1/diff")
 
-        self.assertEqual(resp.status_code, 200)
+        assert resp.status_code == 200
         data = resp.get_json()
-        self.assertTrue(data["success"])
-        self.assertEqual(mock_gh.get_commit_diff.call_count, 1)
+        assert data["success"]
+        assert mock_gh.get_commit_diff.call_count == 1
 
 
-class TestWorkflowListStatusFilter(unittest.TestCase):
+class TestWorkflowListStatusFilter:
     """Tests for workflow list status filter query parameter."""
 
-    def setUp(self):
-        self.client, self.db_path, self.orig, self.db_mod = _make_client()
-
-    def tearDown(self):
-        self.db_mod.adapt_sql = self.orig
-        try:
-            os.unlink(self.db_path)
-        except OSError:
-            pass
-
     @patch("app.routes.autonomous.auto_repo")
-    def test_list_with_status_filter(self, mock_repo):
+    def test_list_with_status_filter(self, mock_repo, client):
         """Status filter is passed through to repo."""
         mock_repo.list_workflows.return_value = []
         mock_repo.count_workflows.return_value = 0
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows?status=completed")
-        self.assertEqual(resp.status_code, 200)
+            resp = client.get("/api/autonomous/workflows?status=completed")
+        assert resp.status_code == 200
         mock_repo.list_workflows.assert_called_once()
 
     @patch("app.routes.autonomous.auto_repo")
-    def test_list_without_status_filter(self, mock_repo):
+    def test_list_without_status_filter(self, mock_repo, client):
         """List all workflows when no status filter provided."""
         mock_repo.list_workflows.return_value = []
         mock_repo.count_workflows.return_value = 0
         with _mock_auth():
-            resp = self.client.get("/api/autonomous/workflows")
-        self.assertEqual(resp.status_code, 200)
+            resp = client.get("/api/autonomous/workflows")
+        assert resp.status_code == 200
         mock_repo.list_workflows.assert_called_once()
-
-
-if __name__ == "__main__":
-    unittest.main()
