--- tests/issues/740/__init__.py
@@ -1,1 +1,0 @@
-"""Batch 4 test package."""
