+++ tests/issues/legacy-directories.txt
@@ -58,9 +58,7 @@
 394
 435
 517
-740
 824
 863
-993
 1006
 1425
