+++ tests/unit/test_autonomous_workflow_idempotency.py
@@ -1,18 +1,22 @@
-"""Tests for Issue #740 Batch 2 — Idempotency, path validation, timeout config, retry limit.
+"""Tests for Issue #740 Batch 2 — Idempotency, timeout config, retry plumbing (unit half).
+
+Migrated from tests/issues/740/test_batch2_idempotency_validation.py. The HTTP
+path-validation and retry-limit route tests moved to
+tests/integration/routes/test_autonomous_path_retry_740.py.
 
 Covers:
 - Idempotency: _find_existing_milestone prevents duplicate milestone creation
-- Path validation: rejects traversal, relative paths, unsafe chars
 - Timeout: configurable via env var and per-workflow task_timeout field
-- Retry limit: max 5 retries for failed workflows
+- _run_agent avoids extra DB queries (uses the passed wf dict)
 """
 
-import os
 from unittest.mock import MagicMock, patch
 
 import pytest
 
 from app.modules.workspace.autonomous.models import AgentTaskResult
+
+pytestmark = [pytest.mark.regression, pytest.mark.issue(740)]
 
 # ── Helpers ──────────────────────────────────────────────────────────
 
@@ -250,206 +254,6 @@ class TestIdempotency:
         mock_repo.create_milestone.assert_called_once()
 
 
-# ── Test: Path Validation ────────────────────────────────────────────
-
-
-class TestPathValidation:
-    """Verify project_path validation in create_workflow."""
-
-    @pytest.fixture(autouse=True)
-    def _allow_quota(self):
-        """These tests exercise path validation, not the quota/rate gate. Stub
-        QuotaManager to allow-by-default so the (real, DB-backed) quota check
-        doesn't reach the test's schema-less DB and spuriously 429 before path
-        validation runs, reset the module-global workflow rate limiter
-        (``_workflow_rate_limiter`` accumulates per-user hits across the whole
-        pytest session, so without a reset these create-workflow requests are
-        429'd once earlier tests in the shard exhaust the 10/hour budget), and
-        stub the module-level ``user_repo.get_user_by_id`` (its db binds at
-        import to a CI db with no users table, so the route's user lookup past
-        path validation 500s once the reset lets valid-path requests through).
-        The #2457 assert-429 / assert-500 cluster."""
-        from app.routes.autonomous import _workflow_rate_limiter
-
-        _workflow_rate_limiter._hits.clear()
-        qmock = MagicMock()
-        qmock.return_value.check_quota.return_value = {"allowed": True, "reason": None}
-        with (
-            patch("app.modules.governance.quota_manager.QuotaManager", qmock),
-            patch(
-                "app.routes.autonomous.user_repo.get_user_by_id",
-                return_value={
-                    "id": 1,
-                    "username": "admin",
-                    "system_account": "",
-                    "role": "admin",
-                    "tenant_id": None,
-                },
-            ),
-        ):
-            yield
-
-    def _make_client(self):
-        """Create Flask test client with test DB."""
-        import tempfile
-
-        import app.repositories.database as db_mod
-
-        db_path = tempfile.mktemp(suffix=".db")
-        orig = db_mod.adapt_sql
-        db_mod.adapt_sql = lambda sql: sql
-
-        db = db_mod.Database(f"sqlite:///{db_path}")
-        try:
-            with db.get_connection() as conn:
-                cursor = conn.cursor()
-                from app.repositories.schema_init import load_schema_from_file
-
-                # Let the authoritative schema create users: the old hand-rolled
-                # CREATE TABLE drifted (no deleted_at/system_account columns —
-                # the schema's partial indexes on those columns then failed).
-                load_schema_from_file(db_url=f"sqlite:///{db_path}", dialect="sqlite")
-                cursor.execute(
-                    "INSERT OR IGNORE INTO users (username, email, password_hash, role) VALUES (?, ?, ?, ?)",
-                    ("admin", "admin@test.com", "hash123", "admin"),
-                )
-                conn.commit()
-        finally:
-            pass
-
-        from app import create_app
-
-        app = create_app({"TESTING": True})
-        c = app.test_client()
-        c.set_cookie("session_token", "test-token")
-        return c, db_path, orig, db_mod
-
-    def _mock_auth(self, user_id=1, role="admin"):
-        return patch(
-            "app.auth.decorators._load_user_from_token",
-            return_value={
-                "id": user_id,
-                "username": "admin" if role == "admin" else "testuser",
-                "email": f"{role}@test.com",
-                "role": role,
-            },
-        )
-
-    def test_rejects_path_traversal(self):
-        """Should reject paths with '..' traversal."""
-        c, db_path, orig, db_mod = self._make_client()
-        repo = MagicMock()
-        repo.create_workflow.return_value = {"workflow_id": "wf-1"}
-
-        with self._mock_auth():
-            with patch("app.routes.autonomous.auto_repo", repo):
-                resp = c.post(
-                    "/api/autonomous/workflows",
-                    json={
-                        "requirements_text": "test",
-                        "cli_tool": "claude-code",
-                        "project_path": "/tmp/../../etc/passwd",
-                    },
-                )
-
-        assert resp.status_code == 400
-        data = resp.get_json()
-        assert "path traversal" in data["error"].lower()
-
-        try:
-            db_mod.adapt_sql = orig
-            os.unlink(db_path)
-        except OSError:
-            pass
-
-    def test_rejects_relative_path(self):
-        """Should reject non-absolute paths."""
-        c, db_path, orig, db_mod = self._make_client()
-        repo = MagicMock()
-        repo.create_workflow.return_value = {"workflow_id": "wf-1"}
-
-        with self._mock_auth():
-            with patch("app.routes.autonomous.auto_repo", repo):
-                resp = c.post(
-                    "/api/autonomous/workflows",
-                    json={
-                        "requirements_text": "test",
-                        "cli_tool": "claude-code",
-                        "project_path": "relative/path/project",
-                    },
-                )
-
-        assert resp.status_code == 400
-        data = resp.get_json()
-        assert "absolute" in data["error"].lower()
-
-        try:
-            db_mod.adapt_sql = orig
-            os.unlink(db_path)
-        except OSError:
-            pass
-
-    def test_accepts_valid_absolute_path(self):
-        """Should accept valid absolute paths."""
-        c, db_path, orig, db_mod = self._make_client()
-        repo = MagicMock()
-        repo.create_workflow.return_value = {
-            "workflow_id": "wf-ok",
-            "title": "test",
-            "status": "pending",
-            "cli_tool": "claude-code",
-        }
-
-        with self._mock_auth():
-            with patch("app.routes.autonomous.auto_repo", repo):
-                resp = c.post(
-                    "/api/autonomous/workflows",
-                    json={
-                        "requirements_text": "test",
-                        "cli_tool": "claude-code",
-                        "project_path": "/home/user/projects/my-app",
-                    },
-                )
-
-        assert resp.status_code == 201
-
-        try:
-            db_mod.adapt_sql = orig
-            os.unlink(db_path)
-        except OSError:
-            pass
-
-    def test_accepts_new_project_without_path(self):
-        """Should accept when is_new_project=true (no path needed)."""
-        c, db_path, orig, db_mod = self._make_client()
-        repo = MagicMock()
-        repo.create_workflow.return_value = {
-            "workflow_id": "wf-new",
-            "title": "new proj",
-            "status": "pending",
-            "cli_tool": "claude-code",
-        }
-
-        with self._mock_auth():
-            with patch("app.routes.autonomous.auto_repo", repo):
-                resp = c.post(
-                    "/api/autonomous/workflows",
-                    json={
-                        "requirements_text": "test",
-                        "cli_tool": "claude-code",
-                        "is_new_project": True,
-                    },
-                )
-
-        assert resp.status_code == 201
-
-        try:
-            db_mod.adapt_sql = orig
-            os.unlink(db_path)
-        except OSError:
-            pass
-
-
 # ── Test: Timeout Configuration ──────────────────────────────────────
 
 
@@ -512,160 +316,6 @@ class TestTimeoutConfig:
 
         call_kwargs = orch._runner.run_agent_task.call_args[1]
         assert call_kwargs["timeout"] == 600
-
-
-# ── Test: Retry Limit ────────────────────────────────────────────────
-
-
-class TestRetryLimit:
-    """Verify retry count limit for failed workflows."""
-
-    def _make_client(self):
-        """Create Flask test client with test DB."""
-        import tempfile
-
-        import app.repositories.database as db_mod
-
-        db_path = tempfile.mktemp(suffix=".db")
-        orig = db_mod.adapt_sql
-        db_mod.adapt_sql = lambda sql: sql
-
-        db = db_mod.Database(f"sqlite:///{db_path}")
-        try:
-            with db.get_connection() as conn:
-                cursor = conn.cursor()
-                from app.repositories.schema_init import load_schema_from_file
-
-                # Let the authoritative schema create users: the old hand-rolled
-                # CREATE TABLE drifted (no deleted_at/system_account columns —
-                # the schema's partial indexes on those columns then failed).
-                load_schema_from_file(db_url=f"sqlite:///{db_path}", dialect="sqlite")
-                cursor.execute(
-                    "INSERT OR IGNORE INTO users (username, email, password_hash, role) VALUES (?, ?, ?, ?)",
-                    ("admin", "admin@test.com", "hash123", "admin"),
-                )
-                conn.commit()
-        finally:
-            pass
-
-        from app import create_app
-
-        app = create_app({"TESTING": True})
-        c = app.test_client()
-        c.set_cookie("session_token", "test-token")
-        return c, db_path, orig, db_mod
-
-    def _mock_auth(self, user_id=1, role="admin"):
-        return patch(
-            "app.auth.decorators._load_user_from_token",
-            return_value={
-                "id": user_id,
-                "username": "admin",
-                "email": f"{role}@test.com",
-                "role": role,
-            },
-        )
-
-    def test_retry_increments_count(self):
-        """Retry should increment retry_count."""
-        c, db_path, orig, db_mod = self._make_client()
-        repo = MagicMock()
-        repo.get_workflow.return_value = {
-            "workflow_id": "wf-retry",
-            "user_id": 1,
-            "status": "failed",
-            "retry_count": 2,
-        }
-        repo.update_workflow.return_value = None
-
-        with self._mock_auth():
-            with patch("app.routes.autonomous.auto_repo", repo):
-                resp = c.post("/api/autonomous/workflows/wf-retry/retry")
-
-        assert resp.status_code == 200
-        call_args = repo.update_workflow.call_args[0]
-        assert call_args[1]["retry_count"] == 3
-
-        try:
-            db_mod.adapt_sql = orig
-            os.unlink(db_path)
-        except OSError:
-            pass
-
-    def test_rejects_retry_over_limit(self):
-        """Should reject retry when retry_count >= MAX_RETRY_COUNT."""
-        c, db_path, orig, db_mod = self._make_client()
-        repo = MagicMock()
-        repo.get_workflow.return_value = {
-            "workflow_id": "wf-max-retry",
-            "user_id": 1,
-            "status": "failed",
-            "retry_count": 5,
-        }
-
-        with self._mock_auth():
-            with patch("app.routes.autonomous.auto_repo", repo):
-                resp = c.post("/api/autonomous/workflows/wf-max-retry/retry")
-
-        assert resp.status_code == 400
-        data = resp.get_json()
-        assert "retry count" in data["error"].lower()
-
-        try:
-            db_mod.adapt_sql = orig
-            os.unlink(db_path)
-        except OSError:
-            pass
-
-    def test_allows_retry_under_limit(self):
-        """Should allow retry when retry_count < MAX_RETRY_COUNT."""
-        c, db_path, orig, db_mod = self._make_client()
-        repo = MagicMock()
-        repo.get_workflow.return_value = {
-            "workflow_id": "wf-under-limit",
-            "user_id": 1,
-            "status": "failed",
-            "retry_count": 4,
-        }
-        repo.update_workflow.return_value = None
-
-        with self._mock_auth():
-            with patch("app.routes.autonomous.auto_repo", repo):
-                resp = c.post("/api/autonomous/workflows/wf-under-limit/retry")
-
-        assert resp.status_code == 200
-
-        try:
-            db_mod.adapt_sql = orig
-            os.unlink(db_path)
-        except OSError:
-            pass
-
-    def test_handles_none_retry_count(self):
-        """Should treat None retry_count as 0."""
-        c, db_path, orig, db_mod = self._make_client()
-        repo = MagicMock()
-        repo.get_workflow.return_value = {
-            "workflow_id": "wf-no-retry",
-            "user_id": 1,
-            "status": "failed",
-            "retry_count": None,
-        }
-        repo.update_workflow.return_value = None
-
-        with self._mock_auth():
-            with patch("app.routes.autonomous.auto_repo", repo):
-                resp = c.post("/api/autonomous/workflows/wf-no-retry/retry")
-
-        assert resp.status_code == 200
-        call_args = repo.update_workflow.call_args[0]
-        assert call_args[1]["retry_count"] == 1
-
-        try:
-            db_mod.adapt_sql = orig
-            os.unlink(db_path)
-        except OSError:
-            pass
 
 
 # ── Test: Idempotency in _create_milestone (wired) ────────────────────
