+++ tests/unit/test_ci_runner.py
@@ -220,12 +220,13 @@ def test_collection_file_count_honors_default_quarantine():
     assert ci.candidate_test_file_count("tests") >= 250
     # Floor mirrors .test-baseline.json issues.min_files (430 -> 413 -> 393 -> 383
     # -> 371 -> 357 -> 338 -> 316 -> 277 -> 244 -> 221 -> 207 -> 192 -> 155
-    # -> 126 -> 110 as #2429 batches 1-8 migrated 17 + 20 + 12 + 19 + 22 + 39 files,
+    # -> 126 -> 110 -> 98 -> 93 as #2429 batches 1-8 migrated 17 + 20 + 12 + 19 + 22 + 39 files,
     # batches 9/10/11/12/13/14 (I-group db, db+subprocess, subprocess-core,
     # merge/gateway/org-sync/run-as, closure, B-group in-proc wave 1)
-    # 33 + 23 + 14 + 15 + 37 + 29 + 16 more into canonical layers / e2e and
-    # governed the stubs (batch 15 = B-group 716/610).
-    assert ci.candidate_test_file_count("tests/issues") >= 110
+    # 33 + 23 + 14 + 15 + 37 + 29 + 16 + 17 more into canonical layers / e2e
+    # and governed the stubs (batch 15 = B-group 716/610, batch 16 = in-proc
+    # closure). min_tests delta 99-143 < 0 -> clamped to remaining count 169.
+    assert ci.candidate_test_file_count("tests/issues") >= 93
 
 
 def test_missing_push_base_fails_safe_to_policy_change():
